0101-0200/0117_populating-next-right-pointers-in-each-node-ii/populating-next-right-pointers-in-each-node-ii.go
package main

/*
  给定一个二叉树
    struct Node {
      int val;
      Node *left;
      Node *right;
      Node *next;
    }
  填充它的每个 next 指针，让这个指针指向其下一个右侧节点。如果找不到下一个右侧节点，则将 next 指针设置为 NULL。
  初始状态下，所有 next 指针都被设置为 NULL。

  进阶：
    你只能使用常量级额外空间。
    使用递归解题也符合要求，本题中递归程序占用的栈空间不算做额外的空间复杂度。

  提示：
    树中的节点数小于 6000
    -100 <= node.val <= 100

  来源：力扣（LeetCode）
  链接：https://leetcode-cn.com/problems/populating-next-right-pointers-in-each-node-ii
  著作权归领扣网络所有。商业转载请联系官方授权，非商业转载请注明出处。
*/

func connect(root *Node) *Node {
	if root == nil {
		return root
	}
	q := []*Node{root}
	for len(q) != 0 {
		curSize := len(q)
		for i := 0; i < curSize; i++ {
			if i > 0 {
				q[i-1].Next = q[i]
			}
			if q[i].Left != nil {
				q = append(q, q[i].Left)
			}
			if q[i].Right != nil {
				q = append(q, q[i].Right)
			}
		}
		q = q[curSize:]
	}
	return root
}

type Node struct {
	Val   int
	Left  *Node
	Right *Node
	Next  *Node
}
